backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0
aiofiles>=23.2.1
musicbrainzngs>=0.7.1
yt-dlp>=2024.1.0
librosa>=0.10.1
pydub>=0.25.1
pyloudnorm>=0.1.1
numpy>=1.24.0
scipy>=1.11.0
motor>=3.3.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
pydantic[email]>=2.0.0
